+++ services/web/src/data/courses/python/syllabus.ts
@@ -1,6 +1,0 @@
-import { Api } from '@game-guild/apiclient';
-import ChapterEntity = Api.ChapterEntity;
-import CourseEntity = Api.CourseEntity;
-import ImageEntity = Api.ImageEntity;
-import UserEntity = Api.UserEntity;
-import LectureEntity = Api.LectureEntity;